Better database (#3)
* First step refacto database

* Refacto database

* Creation ArtworkRepository

* Dispatcher IO ArtworkRepository

* Handle status change

* backup

* Prepare save time

* WatchTime

* Launch player

* Save time

* Rename files

* Delete files

* Better sync

* Fix sync

---------

+++ app/src/main/java/com/kaem/flux/data/tmdb/TMDBService.kt
@@ -23,7 +23,7 @@ interface TMDBService {
 
     @GET("movie/{id}")
     suspend fun getMovieDetails(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Query("language") language: String = Constants.Global.LANGUAGE
     ) : TMDBMovie
 
@@ -36,7 +36,7 @@ interface TMDBService {
 
     @GET("tv/{id}/season/{season}/episode/{episode}")
     suspend fun getEpisode(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Path("season") season: Int,
         @Path("episode") episode: Int,
         @Query("language") language: String = Constants.Global.LANGUAGE
